Add support for pivot table report
This adds a new kind of report that uses the jQuery plugin Cynteka Pivot
Table.
This plugin was inside the source tree and was removed, but added as a
bower dependency.

Use a new service for drop zones

The drop zones of the report creation interface used a lot of similar
functions, and the templates were made of repetitive code. This commit
creates a service for a drop zone, and removes the redundancy by doing
with a single function what was done by 5 different functions.