Add stacking bars functionnality
When there are two dimension fields for a graph, the second field will be used
as a vertical dimension for stacked bars.
In order to make this work, it was necessary to reformat the data,
since the way the c3 plugin creates stacked bars is not compatible with the format of the data
returned by the query.