Redo and restructure the prompt directive
The prompt directive had been partially broken by the report restructuration.
This commit redoes most of the directive, simplifying it and making it work with
the new way data is fetched. In addition, suggested values should be available faster
and be more relevant

This commit does not fix the prompts in dashboards. Minor tweaks are still necessary to make those work.