Remove MongoDB support
The place of MongoDB in widestage is somewhat puzzling.
Some of the code and comments seem to suggest that widestage was
originally intended as a tool to survey MongoDB databases, and that SQL
support was added later.
However, the entire structure of the app seems built around the logic of
SQL databases.
This is most obvious in the layer interface, which is built around
tables with a fixed structures, and joins between these tables (the code
calls these tables 'collections', but uses them exactly as SQL tables).
MongoDB support is acheived through clumsy backend code made to treat a
mongoDB collection like a SQL table (this includes a javascript function
to make a 'join' between the data returned from mongoDB collections).

The result is strange, incomplete, and ultimately unusable. Layers are
built around SQL columns, and MongoDB collections do not have columns.
The code handles this by fetching 100 elements from the collection,
reading all of the surface fields, and treating them like column. This
approach gives access to only some of the data in the table, based on
the arbitrary decisions of how it should be stored.

It seems that there is no simple way to make MongoDB work properly in
urungi. The only good solution would be to redo an entirely new layer
interface, designed for NoSQL databases, and to code the corresponding
backend resolution.

In the meantime, the current mongoDB code is not viable. It never worked
satisfyingly in the first place, and is now completely broken after the
various query restructures. Fixing it would not be worth it.

From now on, and until someone codes a proper NoSQL interface, Urungi is
a SQL data preview app.

Farewell MongoDB, you will be missed.

PS : Urungi still uses MongoDB for it's own database, so don't worry we
won't miss it too much.