Allow to preview only a subset of results
When working with large datasets, editing the report can be tedious
because of the time needed to execute the query.
This patch improves that by making it possible to preview only a subset
of results.
This is done by using a subquery that will fetch a limited number of
lines. The row grouping and ordering is then done on the results of this
subquery.

The query will look like this:

  SELECT <columns...>
  FROM (
    SELECT <columns...>
    FROM <joins...>
    WHERE <conditions...>
    LIMIT <limit>
  ) sub
  GROUP BY <columns...>
  ORDER BY <sorts...>

<limit> can be configured by changing the 'Quick display with' input

This behaviour can be disabled by emptying the 'Quick display with'
input, or setting its value to 0.