Fix layers integrity and remove useless properties
In some cases, it was possible to create custom elements without the
right elements in `arguments` property, resulting in a broken layer
object.

In other cases, it was possible to have some collections and elements
without correct component number.

This patch adds a database migration script that will:
- recalculate component number of all layers,
- delete the `arguments` property of custom elements (this is now
  computed on-the-fly when needed, by parsing the custom element
  expression),
- delete other useless properties which should have never been there in
  the first place (`tempArguments`, `customElements`)

and it fixes the initial problem by removing all occurrences of
`arguments` in code