Do not save reports and dashboards in public space
Reports and dashboards are not supposed to be displayed in the public
space modification page. It was causing reports and dashboard to be
displayed more than once.

This patch fixes the issue by removing reports and dashboards from
public space modification page.
It also refactors some code in dashboards and reports controllers and
makes sure userObjects are always reloaded when needed.

The bug was introduced by 1fd11c39ee23214d14221d43a0606cd8164069d6