Prevent svg chart to overflow its container
c3js determines the chart width based on the container width, but when
the chart is drawn, it increases the page height and can make a
scrollbar appear. This has the effect of reducing the width of all chart
containers without re-calculating the chart width.

This patch fixes this issue by redrawing the chart each time it
overflows its container.

Also, this patch adds the possibility to unset the height of blocks that
were previously set using the inspector.