Add ability to make reports and dashboards public
Public means they are available to anonymous users.

There is a similar feature that allows to publish reports and dashboards
to a company's public space, which is only available to authenticated
users.
To avoid confusion, this has been reworded to "shared space".

Add new dependency on ngclipboard, to copy public reports and dashboards
URLs into the user's clipboard.

Move the 'Share' button to the list of reports and dashboards

Add new AngularJS module 'app.reports'
Add new AngularJS module 'app.dashboards'
Transform AngularJS directive 'ndPrompt' into a component, and rename it
to 'appFilterPrompt'

Update angular-ui-bootstrap to 2.5.6

Co-authored-by: Marot Florian <marot.florian@hotmail.com>