Relocate 'datasourceID' property in layers
datasourceID used to be stored in each collection of the layer. Since
we cannot use more than one datasource in a single layer, this is
redundant.

This patch changes that by moving datasourceID to a top-level property
of layer. It also deletes all other datasourceID properties that were
stored all over the place (in dashboard and reports) and define explicit
schema for reports, dashboards and layers to avoid this property to be
reintroduced accidentally (by an import for instance)

It also modifies the layer creation modal to ask for the datasource to
use. So we cannot have a layer without an associated datasource.