Remove objectLabel and originalLabel properties from reports
originalLabel is essentially a backup for elementLabel,
objectLabel is a duplicate of elementLabel, and
elementLabel is just the concatenation of the layer's element label and
the aggregation name

It does not make sense to store all of this in mongo database.

This patch restores elementLabel to the value of originalLabel if it
exists, and remove originalLabel and objectLabel.
The concatenation of label and aggregation for display is done
dynamically and never stored anywhere.