Fix and improve drag and drop in dashboards and reports
Replace angular-draganddrop by a custom implementation because
angular-draganddrop is not maintained anymore and is not flexible enough

Fix a bug where the user could drop an object anywhere in the DOM tree
of drop target nodes. Now dropped objects are added as direct children
of drop target

Add a box-shadow on drop targets when the user drags a valid object over
it. It makes it easier to see where the object will be dropped.

Also fix image upload

Closes: #37
Closes: #116
Closes: #126
Closes: #146