Remove knex
We do not need a full-featured query builder for the simple SELECT
queries we have. It's actually simpler to build SQL "by hand".

A lot of tests are added to ensure that the built SQL query is correct.

Also remove property `conditionType` in report filters. It was always
equals to 'and' and it was not possible to change it, so it was useless.

Fixes #22