Fix report column internal ID generation
The ID could have been wrongly generated when the associated layer
element had a defaultAggregation set and the user removed this
aggregation.
Now the defaultAggregation is not taken into account when generating the
ID. Also the id is generated when a layer element is dropped onto a
report.
This patch includes a DB migration that regenerates all columns ID for
existing reports and dashboards