Fix display of "runtime filter properties" button for Chromium >= 114
Chromium started implementing the popover API, starting with version 114
It makes use of the popover attribute and for elements that were using
it that meant not being shown at all.
The popover attribute was only used in two places (runtime filter
properties button, and inside the modal opened by it) and it was broken
anyway (the popover would not display even on other browsers), so we
simply remove them.

Related links:
* https://developer.chrome.com/blog/introducing-popover-api/
* https://developer.mozilla.org/en-US/docs/Web/API/Popover_API
* https://html.spec.whatwg.org/multipage/popover.html