Replace report-view directive by a custom element
The report-view directive is still needed for now, because they appear
in the dashboard html that is saved in database.

Important change: the new app-reportview custom element uses shadow DOM
which prevents its theme css to leak outside of the context of the
report.
This allow to use multiple themes in a single dashboard.

Report filters are also re-implemented using custom elements.